Improve: 대시 시스템 수정
