+++ ProjectDaNyan/Assets/ProjectDaNyan/Scripts/UI/ControllerUI/Base/JoystickController.cs
@@ -37,16 +37,12 @@ public class JoystickController : FloatingJoystick
             if (_playerState.getPsData() == PlayerState.PSData.onTheRock)
             {
                 _playerState.setPsData(PlayerState.PSData.exitStartFromRock);
-                PlayDashSound();
             }
             else
             {
                 if (_playerStatus.DashCharged > 0)
                 {
-                    player.gameObject.layer = 7;
-                    _playerState.setPsData(PlayerState.PSData.dash);
-                    _playerStatus.DashCharged -= 1;
-                    PlayDashSound();
+                    _playerState.setPsData(PlayerState.PSData.dashStart);
                 }
                 else
                 {
@@ -89,18 +85,5 @@ public class JoystickController : FloatingJoystick
                 isJoystickPositionGoEnd = false;
             }
         }
-    }
-
-    public void PlayDashSound()
-    {
-        if (dashEffectToggle)
-        {
-            _soundEffectController.playStageSoundEffect(0.5f,SoundEffectController.StageSoundTypes.Player_Dash_0);
-        }
-        else
-        {
-            _soundEffectController.playStageSoundEffect(0.5f,SoundEffectController.StageSoundTypes.Player_Dash_1);
-        }
-        dashEffectToggle = !dashEffectToggle;
     }
 }
